separate block and transaction classes in backend
